rest/requirements.txt
numpy
fastapi
uvicorn
mysql-connector-python
python-dotenv